Moved out networking and hardware into their own classes
