+++ .pre-commit-config.yaml
@@ -1,0 +1,8 @@
+repos:
+  - repo: https://github.com/cpp-linter/cpp-linter-hooks
+    rev: v0.4.0  # Use the ref you want to point at
+    hooks:
+      - id: clang-format
+        args: [--style=Mozilla] # Other coding style: LLVM, GNU, Chromium, Microsoft, Google, WebKit.
+      - id: clang-tidy
+        args: [--checks='boost-*,bugprone-*,performance-*,readability-*,portability-*,modernize-*,clang-analyzer-*,cppcoreguidelines-*']
